+++ include/boost/geometry/core/coordinate_system.hpp
@@ -4,6 +4,10 @@
 // Copyright (c) 2008-2012 Barend Gehrels, Amsterdam, the Netherlands.
 // Copyright (c) 2009-2012 Mateusz Loskot, London, UK.
 
+// This file was modified by Oracle on 2020.
+// Modifications copyright (c) 2020, Oracle and/or its affiliates.
+// Contributed and/or modified by Adam Wulkiewicz, on behalf of Oracle
+
 // Parts of Boost.Geometry are redesigned from Geodan's Geographic Library
 // (geolib/GGL), copyright (c) 1995-2010 Geodan, Amsterdam, the Netherlands.
 
@@ -18,7 +22,7 @@
 #include <boost/mpl/assert.hpp>
 
 #include <boost/geometry/core/point_type.hpp>
-#include <boost/geometry/util/bare_type.hpp>
+#include <boost/geometry/util/type_traits_std.hpp>
 
 
 namespace boost { namespace geometry
@@ -67,7 +71,7 @@ namespace core_dispatch
     {
         typedef typename traits::coordinate_system
             <
-                typename geometry::util::bare_type<Point>::type
+                typename detail::remove_cptrref<Point>::type
             >::type type;
     };
 
@@ -89,7 +93,7 @@ struct coordinate_system
     typedef typename core_dispatch::coordinate_system
         <
             typename tag<Geometry>::type,
-            typename geometry::util::bare_type<Geometry>::type
+            typename detail::remove_cptrref<Geometry>::type
         >::type type;
 };
