+++ include/boost/geometry/extensions/algebra/core/coordinate_dimension.hpp
@@ -5,8 +5,8 @@
 // Copyright (c) 2009-2012 Mateusz Loskot, London, UK.
 // Copyright (c) 2013 Adam Wulkiewicz, Lodz, Poland.
 
-// This file was modified by Oracle on 2018.
-// Modifications copyright (c) 2018 Oracle and/or its affiliates.
+// This file was modified by Oracle on 2018-2020.
+// Modifications copyright (c) 2018-2020 Oracle and/or its affiliates.
 // Contributed and/or modified by Adam Wulkiewicz, on behalf of Oracle
 
 // Parts of Boost.Geometry are redesigned from Geodan's Geographic Library
@@ -19,13 +19,10 @@
 #ifndef BOOST_GEOMETRY_EXTENSIONS_ALGEBRA_CORE_COORDINATE_DIMENSION_HPP
 #define BOOST_GEOMETRY_EXTENSIONS_ALGEBRA_CORE_COORDINATE_DIMENSION_HPP
 
-#include <boost/geometry/core/coordinate_dimension.hpp>
-
-#include <boost/geometry/extensions/algebra/core/tags.hpp>
-
-#include <boost/geometry/util/bare_type.hpp>
-
 #include <boost/mpl/assert.hpp>
+
+#include <boost/geometry/core/coordinate_dimension.hpp>
+#include <boost/geometry/extensions/algebra/core/tags.hpp>
 
 namespace boost { namespace geometry {
 
@@ -36,7 +33,7 @@ struct indexed_dimension
 {
      BOOST_MPL_ASSERT_MSG(false,
                           NOT_IMPLEMENTED_FOR_THIS_GEOMETRY_OR_INDEX,
-                          (Geometry, boost::integral_constant<std::size_t, Index>));
+                          (Geometry, std::integral_constant<std::size_t, Index>));
 };
 
 } // namespace traits
@@ -46,12 +43,12 @@ namespace core_dispatch {
 
 template <typename V>
 struct dimension<vector_tag, V>
-    : traits::dimension<typename geometry::util::bare_type<V>::type>
+    : traits::dimension<typename detail::remove_cptrref<V>::type>
 {};
 
 template <typename G>
 struct dimension<quaternion_tag, G>
-    : traits::dimension<typename geometry::util::bare_type<G>::type>
+    : traits::dimension<typename detail::remove_cptrref<G>::type>
 {};
 
 template <typename T, typename G, std::size_t Index>
@@ -59,23 +56,23 @@ struct indexed_dimension
 {
     BOOST_MPL_ASSERT_MSG(false,
                          NOT_IMPLEMENTED_FOR_THIS_GEOMETRY_OR_INDEX,
-                         (G, boost::integral_constant<std::size_t, Index>));
+                         (G, std::integral_constant<std::size_t, Index>));
 };
 
 template <typename G, std::size_t Index>
 struct indexed_dimension<matrix_tag, G, Index>
-    : traits::indexed_dimension<typename geometry::util::bare_type<G>::type, Index>
+    : traits::indexed_dimension<typename detail::remove_cptrref<G>::type, Index>
 {};
 
 
 template <typename G>
 struct dimension<rotation_quaternion_tag, G>
-    : traits::dimension<typename geometry::util::bare_type<G>::type>
+    : traits::dimension<typename detail::remove_cptrref<G>::type>
 {};
 
 template <typename G>
 struct dimension<rotation_matrix_tag, G>
-    : traits::dimension<typename geometry::util::bare_type<G>::type>
+    : traits::dimension<typename detail::remove_cptrref<G>::type>
 {};
 
 } // namespace core_dispatch
